requirements.txt
# Django
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3

# Security
python-dotenv==1.0.1

# Utils
requests==2.32.3
beautifulsoup4==4.12.3
Unidecode==1.3.8


# Code
black==24.8.0
pylint==3.2.6
pylint-django==2.5.5